# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Architectural Portfolio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /machinery:
    get:
      operationId: listMachinery
      tags: [machinery]
      summary: List all machinery
      parameters:
        - name: published
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of machinery
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MachineryItem"
    post:
      operationId: createMachinery
      tags: [machinery]
      summary: Create machinery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MachineryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MachineryItem"
  /machinery/{slug}:
    get:
      operationId: getMachinery
      tags: [machinery]
      summary: Get machinery by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Machinery detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MachineryItem"
        "404":
          description: Not found
  /machinery/{id}/update:
    put:
      operationId: updateMachinery
      tags: [machinery]
      summary: Update machinery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MachineryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MachineryItem"
  /machinery/{id}/publish:
    patch:
      operationId: toggleMachineryPublish
      tags: [machinery]
      summary: Toggle machinery published status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishToggle"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MachineryItem"
  /machinery/{id}/delete:
    delete:
      operationId: deleteMachinery
      tags: [machinery]
      summary: Delete machinery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/password:
    put:
      operationId: changeAdminPassword
      tags: [admin]
      summary: Change admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminPasswordChange"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "401":
          description: Current password incorrect
        "400":
          description: Validation error
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Unauthorized
...
  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Check admin session
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated
components:
  schemas:
...
    PublishToggle:
      type: object
      required: [published]
      properties:
        published:
          type: boolean
    MachineryItem:
      type: object
      required: [id, name, slug, published]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        category:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        longDescription:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        galleryImages:
          type: ["string", "null"]
        year:
          type: ["string", "null"]
        condition:
          type: ["string", "null"]
        published:
          type: boolean
        featured:
          type: boolean
    MachineryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        category:
          type: string
        description:
          type: string
        longDescription:
          type: string
        imageUrl:
          type: string
        galleryImages:
          type: string
        year:
          type: string
        condition:
          type: string
        published:
          type: boolean
        featured:
          type: boolean
    AdminPasswordChange:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 1
    AdminCredentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
    AdminSession:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean
...