# lib/api-zod/src/generated/types/adminPasswordChange.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Architectural Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminPasswordChange {
  /** @minLength 1 */
  currentPassword: string;
  /** @minLength 1 */
  newPassword: string;
}