# lib/api-zod/src/generated/types/machineryItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Architectural Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface MachineryItem {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  longDescription?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  galleryImages?: string | null;
  /** @nullable */
  year?: string | null;
  /** @nullable */
  condition?: string | null;
  published: boolean;
  featured?: boolean;
}